+++ packages/xchain-bitcoin/src/utils.ts
@@ -331,8 +331,8 @@ export const buildTx = async ({
  * @param {BroadcastTxParams} params The transaction broadcast options.
  * @returns {TxHash} The transaction hash.
  */
-export const broadcastTx = async ({ haskoinUrl, txHex }: BroadcastTxParams): Promise<TxHash> => {
-  return await haskoinApi.broadcastTx({ haskoinUrl, txHex })
+export const broadcastTx = async ({ haskoinUrl, txHex, customRequestHeaders }: BroadcastTxParams): Promise<TxHash> => {
+  return await haskoinApi.broadcastTx({ haskoinUrl, txHex, customRequestHeaders })
 }
 
 /**