+++ packages/xchain-evm/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v.0.1.1 (2022-12-24)
+
+## Update
+
+- Bumped dependencies
+
 # v.0.1.0 (2022-10-13)
 
 ## Update