+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v.0.27.3 (2022-11-24)
+
+## Update
+
+- Bump `xchain-client`
+
 # v.0.27.2 (2022-11-08)
 
 ## Update